+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `INFO` log for `TURN` URIs and username from `TURN` credentials.
 
 ### Changed
+- Allow device checker APIs to take devices as input, rather than only MediaDeviceInfo objects.
 
 - Enable SIMD autodetection for Amazon Voice Focus in Chrome 90+.